% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdm.R
\name{center}
\alias{center}
\title{Mean centers timeseries in a 3D array timeseries x nodes x subjects,
i.e. each timeseries of each node and subject has mean of zero.}
\usage{
center(X)
}
\arguments{
\item{X}{3D array with dimensions timeseries x nodes x subjects.}
}
\value{
M 3D array.
}
\description{
Mean centers timeseries in a 3D array timeseries x nodes x subjects,
i.e. each timeseries of each node and subject has mean of zero.
}

