% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm.R
\name{cor2adj}
\alias{cor2adj}
\title{Threshold correlation matrix to match a given number of edges.}
\usage{
cor2adj(R, n)
}
\arguments{
\item{R}{correlation matrix.}

\item{n}{number of edges.}
}
\value{
A thresholded matrix
}
\description{
Threshold correlation matrix to match a given number of edges.
}
