% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdm.R
\name{corTs}
\alias{corTs}
\title{Correlation of time series.}
\usage{
corTs(ts)
}
\arguments{
\item{ts}{a 3D time series time series x nodes x subjects.}
}
\value{
M correlation matrix.
}
\description{
Correlation of time series.
}
