% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm.R
\name{dgm.group}
\alias{dgm.group}
\title{A group is a list containing restructured data from subejcts for easier group analysis.}
\usage{
dgm.group(subj)
}
\arguments{
\item{subj}{a list of subjects.}
}
\value{
group a list.
}
\description{
A group is a list containing restructured data from subejcts for easier group analysis.
}
\examples{
# create some sample data with 200 samples,
# 3 nodes, and 2 subjects
ts = array(rnorm(200*3*2), dim=c(200,3,2))
mysubs=list()
mysubs[[1]]=subject(ts[,,1])
mysubs[[2]]=subject(ts[,,2])
g=dgm.group(mysubs)

}
