% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm.R
\name{diag.delta}
\alias{diag.delta}
\title{Quick diagnostics on delta.}
\usage{
diag.delta(path, id, nodes)
}
\arguments{
\item{path}{path to results files.}

\item{id}{subject identifier.}

\item{nodes}{number of nodes.}
}
\value{
x array node model's delta
}
\description{
Quick diagnostics on delta.
}
