% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdm.R
\name{dlm.filt.rh}
\alias{dlm.filt.rh}
\title{Calculate the filtering distribution for a specified set of parents and a fixed delta.}
\usage{
dlm.filt.rh(Yt, Ft, delta, m0 = 0, CS0 = 3, n0 = 0.001, d0 = 0.001)
}
\arguments{
\item{Yt}{the vector of observed time series, length T}

\item{Ft}{the matrix of covariates, dim = number of thetas (p) x number of time points (T), usually a column of 1s to represent an intercept and the time series of the parent nodes.}

\item{delta}{discount factor (scalar).}

\item{m0}{the vector of the prior mean at time t=0, length p. The default is a non-informative prior with zero mean. (theta0 | y0, phi) ~ N(m0,CS0*phi^-1).}

\item{CS0}{the prior scale matrix at time t=0, dim = p x p. The default is a non-informative prior, 3 * identity matrix.}

\item{n0}{prior hypermarameter of precision phi ~ G(n0/2; d0/2). The default is a non-informative prior, with a value of 0.001. n0 has to be higher than 0.}

\item{d0}{prior hypermarameter of precision phi ~ G(n0/2; d0/2). The default is a non-informative prior, with a value of 0.001.}
}
\value{
mt the vector or matrix of the posterior mean (location parameter), dim = p x T.
Ct the posterior scale matrix, dim = p x p x T, Ct = CSt * S, where S is a point estimate for the observational variance phi^-1. 
CSt the posterior scale matrix, dim = p x p x T, Ct = CSt * S, where S is a point estimate for the observational variance phi^-1.
Rt the prior scale matrix, dim = p x p x T. Rt = RSt * S_{t-1}, where S_{t-1} is a point estimate for the observational variance phi^-1 at the previous time point.
RSt the prior scale matrix, dim = p X p X T. Rt = RSt * S_{t-1}, where S_{t-1} is a point estimate for the observational variance phi^-1 at the previous time point.
nt and dt the vectors of the hyperparameters for the precision phi with length T.
S the vector of the point estimate for the observational variance phi^-1 with length T.
ft the vector of the one-step forecast location parameter with length T.
Qt the vector of the one-step forecast scale parameter with length T.
ets the vector of the standardised residuals with length T
lpl the vector of the Log Predictive Likelihood with length T.
}
\description{
Calculate the filtering distribution for a specified set of parents and a fixed delta.
}

