% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm.R
\name{dlm_smoo}
\alias{dlm_smoo}
\title{Function for DLM with Smoothing for unknown observational and state variances}
\usage{
dlm_smoo(mt, Ct, Rt, nt, dt, Gt = 0)
}
\arguments{
\item{mt}{the vector or matrix of the posterior mean (location parameter), dim = \code{p x T}.}

\item{Ct}{and \code{CSt} the posterior scale matrix \code{C_{t}} is \code{C_{t} = C*_{t} x S_{t}},
with dim = \code{p x p x T}, where \code{S_{t}} is a point estimate for the observation variance
\code{phi^{-1}}}

\item{Rt}{and \code{RSt} the prior scale matrix \code{R_{t}} is \code{R_{t} = R*_{t} x S_{t-1}},
with dim = \code{p x p x T}, where \code{S_{t-1}} is a point estimate for the observation
variance \code{phi^{-1}} at the previous time point.}

\item{nt}{and the vectors of the updated hyperparameters for the precision \code{phi}
with length \code{T}.}

\item{dt}{the vectors of the updated hyperparameters for the precision \code{phi}
with length \code{T}.}

\item{Gt}{the matrix of state equation with dimension: p X p X T. The default is identity matrix block.}
}
\value{
smt the matrix of smoothing posterior mean with dimension p X T
}
\description{
Function for DLM with Smoothing for unknown observational and state variances
}
