% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dlmLplCpp}
\alias{dlmLplCpp}
\title{C++ implementation of the dlm.lpl}
\usage{
dlmLplCpp(Yt_, Ft_, delta, m0_, CS0_, n0, d0)
}
\arguments{
\item{Yt_}{the vector of observed time series}

\item{Ft_}{the matrix of covariates}

\item{delta}{discount factor}

\item{m0_}{the value of the prior mean}

\item{CS0_}{controls the scaling of the prior variance}

\item{n0}{prior hypermarameter}

\item{d0}{prior hypermarameter}
}
\description{
C++ implementation of the dlm.lpl
}
