% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdm.R
\name{exhaustive.search}
\alias{exhaustive.search}
\title{A function for an exhaustive search, calculates the optimum value of the discount factor.}
\usage{
exhaustive.search(Data, node, nbf = 15, delta = seq(0.5, 1, 0.01),
  cpp = TRUE, m0 = 0, CS0 = 3, n0 = 0.001, d0 = 0.001)
}
\arguments{
\item{Data}{Dataset with dimension number of time points Nt x Number of nodes Nn.}

\item{node}{node of interest.}

\item{nbf}{Log Predictive Likelihood will be calculated from (and including) this time point.}

\item{delta}{a vector of potential values for the discount factor.}

\item{cpp}{boolean true (default): fast C++ implementation, false: native R code.}

\item{m0}{the vector of the prior mean at time t=0, length p. The default is a non-informative prior with zero mean. (theta0 | y0, phi) ~ N(m0,CS0*phi^-1).}

\item{CS0}{the prior scale matrix at time t=0, dim = p x p. The default is a non-informative prior, 3 * identity matrix.}

\item{n0}{prior hypermarameter of precision phi ~ G(n0/2; d0/2). The default is a non-informative prior, with a value of 0.001. n0 has to be higher than 0.}

\item{d0}{prior hypermarameter of precision phi ~ G(n0/2; d0/2). The default is a non-informative prior, with a value of 0.001.}
}
\value{
model.store a matrix with the model, LPL and chosen discount factor for all possible models.
runtime an estimate of the run time of the function, using proc.time().
}
\description{
A function for an exhaustive search, calculates the optimum value of the discount factor.
}

