% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdm.R
\name{getAdjacency}
\alias{getAdjacency}
\title{Get adjacency and associated likelihoods (LPL) and disount factros (df) of winning models.}
\usage{
getAdjacency(winner, nodes)
}
\arguments{
\item{winner, }{2D matrix.}

\item{nodes}{number of nodes.}
}
\value{
adj, 2D adjacency matrix.
}
\description{
Get adjacency and associated likelihoods (LPL) and disount factros (df) of winning models.
}

