% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdm.R
\name{getAdjacencyMatrix}
\alias{getAdjacencyMatrix}
\title{Get adjacency matrix from winning models.}
\usage{
getAdjacencyMatrix(winner, nodes)
}
\arguments{
\item{winner, }{2D matrix.}

\item{nodes}{number of nodes.}
}
\value{
adj, 2D adjacency matrix.
}
\description{
Get adjacency matrix from winning models.
}

