% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm.R
\name{getIncompleteNodes}
\alias{getIncompleteNodes}
\title{Checks results and returns job number for incomplete nodes.}
\usage{
getIncompleteNodes(path, ids, Nr, Nn)
}
\arguments{
\item{path}{path to results.}

\item{ids}{subjects ids.}

\item{Nr}{Number of runs.}

\item{Nn}{Number of nodes.}
}
\value{
jobs job numbers
}
\description{
Checks results and returns job number for incomplete nodes.
}
