% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm.R
\name{getModel}
\alias{getModel}
\title{Extract specific parent model with assocated df and ME from complete model space.}
\usage{
getModel(models, parents)
}
\arguments{
\item{models}{a 2D model matrix.}

\item{parents}{a vector with parent nodes.}
}
\value{
mod specific parent model.
}
\description{
Extract specific parent model with assocated df and ME from complete model space.
}
\examples{
data("utestdata")
r=exhaustive.search(myts,3)
# get model with parents 1, 2, and 4.
m=getModel(r$model.store,c(1,2,4))

}
