% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm.R
\name{getModelNr}
\alias{getModelNr}
\title{Get model number from a set of parents.}
\usage{
getModelNr(models, parents)
}
\arguments{
\item{models}{a 2D model matrix.}

\item{parents}{a vector with parent nodes.}
}
\value{
nr model number.
}
\description{
Get model number from a set of parents.
}
