% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdm.R
\name{group}
\alias{group}
\title{A group is a list containing restructured data from subejcts for easier group analysis.}
\usage{
group(subj)
}
\arguments{
\item{subj}{a list of subjects.}
}
\value{
group a list.
}
\description{
A group is a list containing restructured data from subejcts for easier group analysis.
}

