% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm.R
\name{mergeModels}
\alias{mergeModels}
\title{Merges forward and backward model store.}
\usage{
mergeModels(fw, bw)
}
\arguments{
\item{fw}{forward model.}

\item{bw}{backward model.}
}
\value{
m model store.
}
\description{
Merges forward and backward model store.
}
