% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdm.R
\name{patel}
\alias{patel}
\title{Patel.}
\usage{
patel(X, lower = 0.1, upper = 0.9, P = 0.2, bin = 0.75)
}
\arguments{
\item{X}{time x node 2D matrix.}

\item{lower}{percentile cuttoff.}

\item{upper}{percentile cuttoff for 0-1 scaling.}

\item{P}{probability threshold for connection strength kappa.}

\item{bin}{threshold for conversion to binary values.}
}
\value{
PT list with strengths kappa, direction tau, and net structure.
}
\description{
Patel.
}

