% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdm.R
\name{perf}
\alias{perf}
\title{Performance of estimates, such as sensitivity, specificity, and more.}
\usage{
perf(x, xtrue)
}
\arguments{
\item{x}{estimated binary network matrix.}

\item{xtrue, }{true binary network matrix.}
}
\value{
perf vector.
}
\description{
Performance of estimates, such as sensitivity, specificity, and more.
}

