% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdm.R
\name{perm.test}
\alias{perm.test}
\title{Permutation test for Patel's kappa. Creates a distribution of values
kappa under the null hypothesis.}
\usage{
perm.test(X, alpha = 0.05)
}
\arguments{
\item{X}{time x node x subjects 3D matrix.}

\item{alpha}{sign. level}
}
\value{
stat lower and upper significance thresholds.
}
\description{
Permutation test for Patel's kappa. Creates a distribution of values
kappa under the null hypothesis.
}
