% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm.R
\name{pruning}
\alias{pruning}
\title{Get pruned adjacency network.}
\usage{
pruning(adj, models, winner, e = 20)
}
\arguments{
\item{adj}{list with network adjacency from getAdjacency().}

\item{models}{matrix 3D with full model estimates.}

\item{winner}{matrix 2D with winning models.}

\item{e}{bayes factor for network pruning.}
}
\value{
thr list with pruned network adjacency.
}
\description{
Get pruned adjacency network.
}
