% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdm.R
\name{read.subject}
\alias{read.subject}
\title{Reads single subject's network from text files.}
\usage{
read.subject(id, nodes)
}
\arguments{
\item{id}{subject ID.}

\item{nodes}{number of nodes.}
}
\value{
store list with results.
}
\description{
Reads single subject's network from text files.
}

