% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm.R
\name{reshapeTs}
\alias{reshapeTs}
\title{Reshapes a 2D concatenated time series into 3D according to no. of subjects and volumes.}
\usage{
reshapeTs(ts, N, V)
}
\arguments{
\item{ts}{a 2D time series volumes x nodes.}

\item{N}{No. of subjects.}

\item{V}{No. of volumes.}
}
\value{
M 3D matrix, time series x nodes x subjects.
}
\description{
Reshapes a 2D concatenated time series into 3D according to no. of subjects and volumes.
}
