% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm.R
\name{rmdiag}
\alias{rmdiag}
\title{Removes diagnoal from matrix with NAs.}
\usage{
rmdiag(M)
}
\arguments{
\item{M}{Matrix}
}
\value{
matrix with diagnoal of NAs.
}
\description{
Removes diagnoal from matrix with NAs.
}
