% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm.R
\name{rmna}
\alias{rmna}
\title{Removes NAs from matrix.}
\usage{
rmna(M)
}
\arguments{
\item{M}{Matrix}
}
\value{
matrix with NAs removed.
}
\description{
Removes NAs from matrix.
}
\examples{
M=array(NA, dim=c(3,3))
M[1,2]=0.9
M=rmna(M)
}
