% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm.R
\name{rmRecipLow}
\alias{rmRecipLow}
\title{Removes reciprocal connections in the lower diagnoal of the network matrix.}
\usage{
rmRecipLow(M)
}
\arguments{
\item{M}{adjacency matrix}
}
\value{
M adjacency matrix without reciprocal connections.
}
\description{
Removes reciprocal connections in the lower diagnoal of the network matrix.
}
