% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdm.R
\name{scaleTs}
\alias{scaleTs}
\title{Scaling data. Zero centers and scales the nodes (SD=1).}
\usage{
scaleTs(X)
}
\arguments{
\item{X}{time x node 2D matrix, or 3D with subjects as the 3rd dimension.}
}
\value{
S centered and scaled matrix.
}
\description{
Scaling data. Zero centers and scales the nodes (SD=1).
}
