% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdm.R
\name{stepwise.combine}
\alias{stepwise.combine}
\title{Stepise combine: combines the stepwise forward and the stepwise backward model.}
\usage{
stepwise.combine(forward_matrix, backward_matrix)
}
\arguments{
\item{forward_matrix}{The winning sets of parents using a Forward Selection model search. A
matrix with dimension \code{Nn+2} x \code{Nn}, rows \code{1:Nn} are the parents (ones and zeros),
rows \code{(Nn+1):(Nn+2)} are the LPL and discount factor. forward matrix.}

\item{backward_matrix}{backward_matrix}{The winning sets of parents using a Backward Elimination
model search. A matrix with dimension \code{Nn+2} x \code{Nn}, rows \code{1:Nn} are the parents
(ones and zeros), rows \code{(Nn+1):(Nn+2)} are the LPL and discount factor.}
}
\value{
stepwise_combine_matrix The adjacency network, LPLs and discount factors when the Forward
Selection and Backward Elimination model searches are combined.
}
\description{
Stepise combine: combines the stepwise forward and the stepwise backward model.
}
