% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm.R
\name{symmetric}
\alias{symmetric}
\title{Turns asymetric network into an symmetric network. Helper function to
determine the detection of a connection while ignoring directionality.}
\usage{
symmetric(M)
}
\arguments{
\item{M}{3D matrix nodes x nodes x subjects}
}
\value{
3D matrix nodes x nodes x subjects
}
\description{
Turns asymetric network into an symmetric network. Helper function to
determine the detection of a connection while ignoring directionality.
}
\examples{
M=array(NA, dim=c(3,3,2))
M[,,1]=matrix(c(0,0,0,1,0,0,0,1,0),3,3)
M[,,2]=matrix(c(0,0,0,1,0,0,0,0,0),3,3)
M_=symmetric(M)
}
