% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm.R
\name{ttest.nettest}
\alias{ttest.nettest}
\title{Comparing connectivity strenght of two groups with FDR correction.}
\usage{
ttest.nettest(s, g, alpha = 0.05, fdr = 0.05)
}
\arguments{
\item{s}{matrix with Nn x Nn x N.}

\item{g}{group assignment, vector of type factor of size N.}

\item{alpha}{alpha level for uncorrected test.}

\item{fdr}{FDR alpha level.}
}
\value{
store List with test statistics and p-values.
}
\description{
Comparing connectivity strenght of two groups with FDR correction.
}
