% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdm.R
\name{center}
\alias{center}
\title{Mean centers timeseries in a 2D array timeseries x nodes,
i.e. each timeseries of each node has mean of zero.}
\usage{
center(X)
}
\arguments{
\item{X}{2D array with dimensions timeseries x nodes.}
}
\value{
M 2D array.
}
\description{
Mean centers timeseries in a 2D array timeseries x nodes,
i.e. each timeseries of each node has mean of zero.
}

