% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm.R
\name{corTs}
\alias{corTs}
\title{Mean correlation of time series across subjects.}
\usage{
corTs(ts)
}
\arguments{
\item{ts}{a 3D time series time series x nodes x subjects.}
}
\value{
M correlation matrix.
}
\description{
Mean correlation of time series across subjects.
}
\examples{
# create some sample data with 200 samples,
# 5 nodes, and 2 subjects
ts = array(rnorm(200*5*2), dim=c(200,5,2))
M = corTs(ts)

}
