% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm.R
\name{dgm.group}
\alias{dgm.group}
\title{A group is a list containing restructured data from subejcts for easier group analysis.}
\usage{
dgm.group(subj, saveModels = TRUE)
}
\arguments{
\item{subj}{a list of subjects.}

\item{saveModels}{whether to save models or not.}
}
\value{
group a list.
}
\description{
A group is a list containing restructured data from subejcts for easier group analysis.
}
