% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdm.R
\name{dlm.filt.rh}
\alias{dlm.filt.rh}
\title{Calculate the filtering distribution for a specified set of parents and a fixed delta.}
\usage{
dlm.filt.rh(Yt, Ft, delta, m0 = numeric(nrow(Ft)), CS0 = 3 * diag(nrow(Ft)),
  n0 = 0.001, d0 = 0.001)
}
\arguments{
\item{Yt}{time series of the node of interest (dim = Nt).}

\item{Ft}{the time series of the parents (number of parents = p), and 1 for the intercept (dim = p x Nt).}

\item{delta}{discount factor (scalar).}

\item{m0}{prior means at time t=0 with length p. The default is non-informative prior, with zero mean.}

\item{CS0}{squared matrix of prior variance. The default is non-informative prior, with prior variance equal to 3 times the observed variance.}

\item{n0}{prior hypermarameters of precision phi ~ G(n0/2; d0/2). The default is non-informative priors, with value of 0.001. n0 has to be higher than 0.}

\item{d0}{prior hypermarameters of precision phi ~ G(n0/2; d0/2). The default is non-informative priors, with value of 0.001. n0 has to be higher than 0.}
}
\value{
mt = the filtered posterior mean, dim = p x T.
Ct = the filtered posterior variance, dim = p x p x T.
CSt.
Rt = the prior variance, dim = p X p X T.
RSt
nt and dt = the prior hypermarameters of precision phi with length T.
ft = the one-step forecast mean with length T.
Qt = the one-step forecast variance with length T.
ets = the standardised residuals with length T.
lpl = Log Predictive Likelihood with length T.
}
\description{
Calculate the filtering distribution for a specified set of parents and a fixed delta.
}

