% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm.R
\name{getIncompleteNodes}
\alias{getIncompleteNodes}
\title{Returns job number for incomplete nodes.}
\usage{
getIncompleteNodes(path)
}
\arguments{
\item{path}{path to results.}
}
\value{
jobs job numbers
}
\description{
Returns job number for incomplete nodes.
}
