% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdm.R
\name{getModel}
\alias{getModel}
\title{Get specific parent model from all models.}
\usage{
getModel(models, parents)
}
\arguments{
\item{models}{a 2D model matrix.}

\item{parents}{a vector with parent nodes.}
}
\value{
mod specific parent model.
}
\description{
Get specific parent model from all models.
}

