% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm.R
\name{getThreshAdj}
\alias{getThreshAdj}
\title{Get thresholded adjacency network.}
\usage{
getThreshAdj(adj, models, winner, bf = 20)
}
\arguments{
\item{adj}{list with network adjacency from getAdjacency().}

\item{models}{matrix 3D with full model estimates.}

\item{winner}{matrix 2D with winning models.}

\item{bf}{bayes factor for network thresholding.}
}
\value{
thr list with thresholded network adjacency.
}
\description{
Get thresholded adjacency network.
}
