% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdm.R
\name{gplotMat}
\alias{gplotMat}
\title{Plots network as adjacency matrix.}
\usage{
gplotMat(adj, title = NULL, label = NULL, hasColMap = NULL, lim = c(0,
  1))
}
\arguments{
\item{adj}{2D adjacency matrix.}

\item{title}{title.}

\item{label}{label for colormap.}

\item{hasColMap}{FALSE turns off color map, default is NULL (on).}

\item{lim}{vector with min and max value for color scaling.}
}
\description{
Plots network as adjacency matrix.
}
