% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdm.R
\name{model.generator}
\alias{model.generator}
\title{A function to generate all the possible models.}
\usage{
model.generator(Nn, node)
}
\arguments{
\item{Nn}{number of nodes; the number of columns of the dataset can be used.}

\item{node}{the node of interest (i.e., the node to find parents for).}
}
\value{
output.model = a matrix with dimensions (nn-1) x number of models, where number of models = 2^(nn-1).
}
\description{
A function to generate all the possible models.
}

