% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdm.R
\name{node}
\alias{node}
\title{Runs exhautive search on a single node. In addition to exhautive.search() 
this function can save results in txt file.}
\usage{
node(X, n, id = NULL)
}
\arguments{
\item{X}{3D array with dimensions timeseries x nodes x subjects.}

\item{n}{node number.}

\item{id}{subject ID. If set, results are saved to a txt file.}
}
\value{
store list with results.
}
\description{
Runs exhautive search on a single node. In addition to exhautive.search() 
this function can save results in txt file.
}

