% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdm.R
\name{node}
\alias{node}
\title{Runs exhaustive search on a single node and saves results in txt file.}
\usage{
node(X, n, id = NULL)
}
\arguments{
\item{X}{array with dimensions timeseries x nodes.}

\item{n}{node number.}

\item{id}{subject ID. If set, results are saved to a txt file.}
}
\value{
store list with results.
}
\description{
Runs exhaustive search on a single node and saves results in txt file.
}

