% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdm.R
\name{patel}
\alias{patel}
\title{Patel.}
\usage{
patel(X, lower = 0.1, upper = 0.9, bin = 0.75, TK = 0, TT = 0)
}
\arguments{
\item{X}{time x node 2D matrix.}

\item{lower}{percentile cuttoff.}

\item{upper}{percentile cuttoff for 0-1 scaling.}

\item{bin}{threshold for conversion to binary values.}

\item{TK}{significance threshold for connection strength kappa.}

\item{TT}{significance threshold for direction tau.}
}
\value{
PT list with strengths kappa, direction tau, and net structure.
}
\description{
Patel.
}

