% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm.R
\name{perf}
\alias{perf}
\title{Performance of estimates, such as sensitivity, specificity, and more.}
\usage{
perf(x, true)
}
\arguments{
\item{x}{estimated binary network matrix.}

\item{true, }{true binary network matrix.}
}
\value{
p list with results.
}
\description{
Performance of estimates, such as sensitivity, specificity, and more.
}
