% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm.R
\name{pruning}
\alias{pruning}
\title{Get pruned adjacency network.}
\usage{
pruning(adj, models, winner, e = 20)
}
\arguments{
\item{adj}{list with network adjacency from getAdjacency().}

\item{models}{list of models.}

\item{winner}{matrix 2D with winning models.}

\item{e}{bayes factor for network pruning.}
}
\value{
thr list with pruned network adjacency.
}
\description{
Get pruned adjacency network.
}
\examples{
data("utestdata")
# select only 3-nodes to speed-up this example
sub=subject(myts[,1:3])
p=pruning(sub$adj, sub$models, sub$winner)
}
