% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm.R
\name{read.subject}
\alias{read.subject}
\title{Reads single subject's network from txt files.}
\usage{
read.subject(path, id, nodes, modelStore = TRUE)
}
\arguments{
\item{path}{path.}

\item{id}{identifier to select all subjects' nodes, e.g. pattern containing subject ID and session number.}

\item{nodes}{number of nodes.}

\item{modelStore}{can be set to false to save memory.}
}
\value{
store list with results.
}
\description{
Reads single subject's network from txt files.
}
