% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm.R
\name{rmdiag}
\alias{rmdiag}
\title{Removes diagonal of NA's from matrix.}
\usage{
rmdiag(M)
}
\arguments{
\item{M}{Matrix}
}
\value{
matrix with diagonal of 0's.
}
\description{
Removes diagonal of NA's from matrix.
}
\examples{
M=array(rnorm(3*3), dim=c(3,3))
M[as.logical(diag(3))] = NA
M=rmna(M)
}
