% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdm.R
\name{subject}
\alias{subject}
\title{Estimate subject's full network: runs exhaustive search on very node.}
\usage{
subject(X, id = NULL)
}
\arguments{
\item{X}{array with dimensions timeseries x nodes.}

\item{id}{subject ID. If set, results are saved to a txt file.}
}
\value{
store list with results.
}
\description{
Estimate subject's full network: runs exhaustive search on very node.
}

