% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdm.R
\name{symmetric}
\alias{symmetric}
\title{Turns asymetric network into an symmetric network. Helper function to
determine the detection of a connection while ignoring directionality.}
\usage{
symmetric(M)
}
\arguments{
\item{M}{3D matrix nodes x nodes x subjects}
}
\value{
3D matrix nodes x nodes x subjects
}
\description{
Turns asymetric network into an symmetric network. Helper function to
determine the detection of a connection while ignoring directionality.
}
