% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgm.R
\name{ttest.nettest}
\alias{ttest.nettest}
\title{Comparing connectivity strenght of two groups with FDR correction.}
\usage{
ttest.nettest(m, g, alpha = 0.05, fdr = 0.05, perm = FALSE,
  n_perm = 9999)
}
\arguments{
\item{m}{matrix with Nn x Nn x N.}

\item{g}{group assignment, vector of type factor of size N.}

\item{alpha}{alpha level for uncorrected test.}

\item{fdr}{FDR alpha level.}

\item{perm}{optional permuation test, default is false.}

\item{n_perm}{number of permutations.}
}
\value{
store List with test statistics and p-values.
}
\description{
Comparing connectivity strenght of two groups with FDR correction.
}
